+++ crux/grafana/main/alertmanager.service
@@ -1,0 +1,20 @@
+[Unit]
+Description=Alertmanager
+Wants=network-online.target
+After=network-online.target
+
+StartLimitIntervalSec=10
+StartLimitBurst=5
+
+[Service]
+User=alertmanager
+Group=alertmanager
+Type=simple
+Restart=always
+RestartSec=5s
+ExecStart=/usr/local/bin/alertmanager \
+  --storage.path=/alertmanager-data \
+  --config.file=/etc/alertmanager/alertmanager.yml
+
+[Install]
+WantedBy=multi-user.target
