+++ crux/grafana/main/alertmanager.yml
@@ -1,0 +1,29 @@
+---
+route:
+  group_by: ['alertname']
+  group_wait: 30s
+  group_interval: 5m
+  repeat_interval: 1h
+  receiver: 'web.hook'
+  routes:
+  - receiver: slack-notifications
+    match:
+      severity: warning
+receivers:
+- name: 'web.hook'
+  webhook_configs:
+  - url: 'http://127.0.0.1:5001/'
+- name: slack-notifications
+  slack_configs:
+  - channel: "#grafana-alert"
+    send_resolved: true
+    api_url: "REPLACE_WITH_WEBHOOK_URL"
+    title: "<@REPLACE_WITH_USERNAME_ID> {{ .GroupLabels.alertname }}"
+    text: "{{ range .Alerts }}{{ .Annotations.message }}\n{{ end }}"
+    icon_url: https://avatars3.githubusercontent.com/u/3380462
+inhibit_rules:
+  - source_match:
+      severity: 'critical'
+    target_match:
+      severity: 'warning'
+    equal: ['alertname', 'dev', 'instance']
