+++ crux/grafana/main/ben-error-rule.yml
@@ -1,0 +1,11 @@
+---
+groups:
+- name: ben-error-rules
+  rules:
+  - alert: NewBenErrorLogs
+    annotations:
+      message: Ben Job Scheduler logs reported new error.
+    expr: delta(ben_errors_count{job="backup"}[15s]) > 0
+    for: 1m
+    labels:
+      severity: error
