+++ crux/grafana/main/ben-grafana.service
@@ -1,0 +1,12 @@
+[Unit]
+Description=Ben Grafana Service
+After=network-online.target
+
+[Service]
+User=ben
+ExecStart=/usr/bin/python3 /etc/ben/node_util.py
+Type=simple
+Restart=on-failure
+
+[Install]
+WantedBy=default.target
