+++ crux/grafana/main/ben-jobs.service
@@ -1,0 +1,12 @@
+[Unit]
+Description=Ben Grafana Service
+After=network-online.target
+
+[Service]
+User=ben
+ExecStart=/usr/bin/python3 /etc/ben/ben-jobs.py
+Type=simple
+Restart=on-failure
+
+[Install]
+WantedBy=default.target
