+++ crux/grafana/main/ben-logs.service
@@ -1,0 +1,12 @@
+[Unit]
+Description=Ben Log Counter for PushGateway
+After=network-online.target
+
+[Service]
+User=ubuntu
+ExecStart=/bin/bash /etc/ben/error_counter.sh
+Type=simple
+Restart=on-failure
+
+[Install]
+WantedBy=default.target
