+++ crux/grafana/main/ben-logs.timer
@@ -1,0 +1,10 @@
+[Unit]
+Description=Ben Log Counter for PushGateway
+
+[Timer]
+OnBootSec=15min
+OnUnitActiveSec=1h
+Persistent=true
+
+[Install]
+WantedBy=timers.target
