+++ crux/grafana/main/loki.service
@@ -1,0 +1,11 @@
+[Unit]
+Description=Loki service
+After=network.target
+
+[Service]
+Type=simple
+User=root
+ExecStart=/usr/local/bin/loki -config.file /etc/loki-local-config.yml
+
+[Install]
+WantedBy=multi-user.target
