+++ crux/grafana/main/pushgateway.service
@@ -1,0 +1,18 @@
+[Unit]
+Description=Pushgateway
+Wants=network-online.target
+After=network-online.target
+
+StartLimitIntervalSec=500
+StartLimitBurst=5
+
+[Service]
+User=pushgateway
+Group=pushgateway
+Type=simple
+Restart=on-failure
+RestartSec=5s
+ExecStart=/usr/local/bin/pushgateway
+
+[Install]
+WantedBy=multi-user.target
