+++ crux/misc/taxa_from_fasta.sh
@@ -19,15 +19,15 @@ while getopts "d:t:f:i:k:s:r:" opt; do
     esac
 done
 
-export AWS_ACCESS_KEY_ID=${AWS_ACCESS_KEY_ID}
-export AWS_SECRET_ACCESS_KEY=${AWS_SECRET_ACCESS_KEY}
-export AWS_DEFAULT_REGION=${AWS_DEFAULT_REGION}
+# export AWS_ACCESS_KEY_ID=${AWS_ACCESS_KEY_ID}
+# export AWS_SECRET_ACCESS_KEY=${AWS_SECRET_ACCESS_KEY}
+# export AWS_DEFAULT_REGION=${AWS_DEFAULT_REGION}
 
 
 primer=$(echo $FOLDER | cut -d"-" -f1)
 # download master taxa and local fasta
 aws s3 cp s3://ednaexplorer/crux/$RUNID/fa-taxid/$primer.tax.tsv . --endpoint-url https://js2.jetstream-cloud.org:8001/
-aws s3 cp s3://ednaexplorer/tronko/$RUNID/$FOLDER/$FASTA . --endpoint-url https://js2.jetstream-cloud.org:8001/
+# aws s3 cp s3://ednaexplorer/tronko/$RUNID/$FOLDER/$FASTA . --endpoint-url https://js2.jetstream-cloud.org:8001/
 
 
 skip="FALSE"
@@ -38,8 +38,7 @@ while read line; do
         accid="${line:1}"
         taxline=$(grep "$accid" $primer.tax.tsv)
         echo $taxline >> $TAXA
-        sed 's/ /\t/' $TAXA > tmp && mv tmp $TAXA # grep converts tabs to space. Converting space back to tabs
     fi
 done < $FASTA # master tax.tsv
-
-aws s3 cp $TAXA s3://ednaexplorer/tronko/$RUNID/$FOLDER/$TAXA --endpoint-url https://js2.jetstream-cloud.org:8001/
+sed 's/ /\t/' $TAXA > tmp && mv tmp $TAXA # grep converts tabs to space. Converting space back to tabs
+# aws s3 cp $TAXA s3://ednaexplorer/tronko/$RUNID/$FOLDER/$TAXA --endpoint-url https://js2.jetstream-cloud.org:8001/
