+++ crux/vars/crux_vars.sh
@@ -3,11 +3,11 @@ CYVERSE="/vars/config.yaml" # cyverse config file name
 CYVERSE_BASE="/iplant/home/shared/eDNA_Explorer/crux"
 BWA_INDEX_URL="https://data.cyverse.org/dav-anon${CYVERSE_BASE}/bwa-index"
 THREADS=16
-BLAST_THREADS=4 # each thread takes about 18GB of RAM
+BLAST_THREADS=4
 RUNID="2023-04-07"
 
 # Obitools ecopcr variables
-ERROR=3
+ERR=3
 # MINLENGTH=30
 # MAXLENGTH=1000
