--- tronko/ben-logs.timer
@@ -1,9 +1,0 @@
-[Unit]
-Description=Ben Log Counter for PushGateway
-
-[Timer]
-OnCalendar=*-*-* */1:00:00
-Persistent=true
-
-[Install]
-WantedBy=timers.target
